+++ package.json
@@ -30,6 +30,7 @@
     "express": "^4.17.1",
     "graphql": "^14.6.0",
     "jsonwebtoken": "^8.5.1",
+    "lodash": "^4.17.15",
     "mongoose": "^5.9.5",
     "query-string": "^6.11.1"
   }
